+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changes
 development version
 -------------------
 
+* The IgBLAST cache is now disabled by default. We assume that, in most cases, datasets will not be
+  re-run with the exact same parameters, and then it only fills up the disk. Delete your cache with
+  ``rm -r ~/.cache/igdiscover`` to reclaim the space.
 * Added detection of chimeras to the (pre-)germline filters. Any novel allele that can be explained
   as a chimera of two unmodified reference alleles is marked in the ``new_V_germline.tab`` file
   (the candidate is currently not discarded).