+++ CHANGES.rst
@@ -14,6 +14,7 @@ development version
   incorrectly removed some candidates that have a mutation close to the 5' end.
   This was replaced with a simpler filter that only ensures that there are no
   two candidates with the same sequence.
+* Use IgBLAST 1.10
 
 v0.11 (2018-11-27)
 ------------------